Code Change to incorporate solr changes with 1.12 changes.
Design: With 1.12 these is a more robust way of using tuque to query, which we need but we also need to use solr as well. So with this based on the form option it will decide whether or not to query based on tuque or solr. This also allows for us to cron updates with solr as well.